migrated httpInterface routes to individual files
