+++ Platform/Client/Http-Routes/chart-layers.js
@@ -1,0 +1,14 @@
+exports.newChartLayersRoute = function newChartLayersRoute() {
+    const thisObject = {
+        endpoint: 'ChartLayers',
+        command: command
+    }
+
+    return thisObject
+
+    function command(httpRequest, httpResponse) {
+        let requestPathAndParameters = httpRequest.url.split('?') // Remove version information
+        let requestPath = requestPathAndParameters[0].split('/')
+        SA.projects.foundations.utilities.httpResponses.respondWithFile(global.env.PATH_TO_PLATFORM + '/UI/' + endpointOrFile + '/' + requestPath[2], httpResponse)
+    }
+}
