+++ Platform/Client/Http-Routes/list-global-files.js
@@ -1,0 +1,12 @@
+exports.newListGlobalFilesRoute = function newListGlobalFilesRoute() {
+    const thisObject = {
+        endpoint: 'ListGlobalFiles',
+        command: command
+    }
+
+    return thisObject
+
+    function command(httpRequest, httpResponse) {
+        SA.projects.foundations.utilities.httpResponses.respondWithProjectFolderFileList(httpResponse, 'Globals', 'UI')
+    }
+}
