+++ Platform/Client/Http-Routes/list-system-action-functions.js
@@ -1,0 +1,12 @@
+exports.newListSystemActionFunctionsRoute = function newListSystemActionFunctionsRoute() {
+    const thisObject = {
+        endpoint: 'ListSystemActionFunctions',
+        command: command
+    }
+
+    return thisObject
+
+    function command(httpRequest, httpResponse) {
+        SA.projects.foundations.utilities.httpResponses.respondWithProjectFolderFileList(httpResponse, 'System-Action-Functions', 'UI')
+    }
+}
