+++ Platform/Client/Http-Routes/list-utilities-files.js
@@ -1,0 +1,12 @@
+exports.newListUtilitiesFilesRoute = function newListUtilitiesFilesRoute() {
+    const thisObject = {
+        endpoint: 'ListUtilitiesFiles',
+        command: command
+    }
+
+    return thisObject
+
+    function command(httpRequest, httpResponse) {
+        SA.projects.foundations.utilities.httpResponses.respondWithProjectFolderFileList(httpResponse, 'Utilities', 'UI')
+    }
+}
